Statistics Tool for Steam in ACP
There is now a statistics tool for Steam Games (count of owned games by those registered on Steam with a public profile).  This follows the same requirements of the Daily Statistics in terms of permissions.  A cron entry is added that updates Steam users twice a day so the statistics do take time to update.