Fixed installer. No more duplicate column error.
Modified Manufacture.php to include function called "addColumnIfNotExists".

This is a function that accepts a tableName, fieldName, fieldDefinition and after parameters.

The function first looks at all columns in the specified table matching the field name. If the column exists, nothing happens (therefore no error).

If the column does not exist then it is created with the specified field definition after the specified field.